pencil-art/requirements.txt
Flask==3.1.2
opencv-python==4.10.0.84
numpy==2.1.3
Pillow==11.0.0
Werkzeug==3.1.3
